% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfrs.R
\name{bfrs}
\alias{bfrs}
\title{Bayesian functional regression model with Stan}
\usage{
bfrs(
  formula,
  data,
  family = gaussian(),
  joint_FPCA = NULL,
  intercept = TRUE,
  n.iter = 5000,
  n.warmup = 2000,
  n.knots = 1,
  cens = NULL,
  func_parameter = NULL,
  runStan = TRUE,
  ...
)
}
\arguments{
\item{formula}{functional regression formula}

\item{data}{data.frame containing data of all variables used in the model}

\item{family}{type of outcome variable, bfrs currently support "guassian", "binomial", "Cox", and "functional"}

\item{joint_FPCA}{A vector of the same length of the functional predictors, indicating whether jointly modeling FPCA for the functional predictors.}

\item{intercept}{True/False variable for whether include an intercept term in the linear predictor}

\item{n.iter}{Total number of Bayesian iterations}

\item{n.warmup}{Number of warmup (burnin) iterations for posterior sampling}

\item{n.knots}{Number of cores to use when executing the chains in parallel}

\item{cens}{Indicator variable for censoring if family = "Cox"}

\item{func_parameter}{Specify the spline parameter for the FoSR.}

\item{runStan}{True/False variable for whether to run the Stan program. If false, only generate the Stan code and data.}
}
\description{
Fit the Bayesian functional regression model using R Stan. At this moment, bfrs supports the Scalar-on-Function Regression (SoFR),
functional Cox regression, and Function-on-Scalar Regression (FoSR).
}
