% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesdata.R
\name{ext_func_pred_obs}
\alias{ext_func_pred_obs}
\title{A function for extracting the data for the functional predictors and organize them as input for Stan}
\usage{
ext_func_pred_obs(term, data)
}
\description{
A function for extracting the data for the functional predictors and organize them as input for Stan
}
