% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesdata.R
\name{ext_scalar_pred}
\alias{ext_scalar_pred}
\title{A function for extracting the data for the scalar predictors and organize them as input for Stan}
\usage{
ext_scalar_pred(term, data)
}
\description{
A function for extracting the data for the scalar predictors and organize them as input for Stan
}
