% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesbasis.R
\name{extract_basis}
\alias{extract_basis}
\title{Functions to extract the spline basis from data based on formula}
\usage{
extract_basis(formula, data, func_comp)
}
\description{
The main function for extracting the spline basis from data based on formula
}
