% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesbasis.R
\name{extract_basis_cc}
\alias{extract_basis_cc}
\title{Extract basis for cyclic cubic regression spline ("cc")}
\usage{
extract_basis_cc(term, data)
}
\description{
Extract basis for cyclic cubic regression spline ("cc")
}
