% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesbasis.R
\name{extract_basis_cr}
\alias{extract_basis_cr}
\title{Extract basis for cubic regression spline ("cr")}
\usage{
extract_basis_cr(term, data)
}
\description{
Extract basis for cubic regression spline ("cr")
}
