% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesformula.R
\name{extract_var}
\alias{extract_var}
\title{A function for identifying the scalar and functional predictors}
\usage{
extract_var(term)
}
\description{
A function for identifying the scalar and functional predictors
}
