% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr_bayes.R
\name{fosr_bayes}
\alias{fosr_bayes}
\title{Bayesian Function-on-Scalar Regression}
\usage{
fosr_bayes(
  formula,
  data,
  joint_FPCA = NULL,
  runStan = TRUE,
  niter = 3000,
  nwarmup = 1000,
  nchain = 3,
  ncores = 1,
  spline_type = "bs",
  spline_df = 10
)
}
\arguments{
\item{formula}{Functional regression formula, with the same syntax as that in the R mgcv package.}

\item{data}{A data frame containing data of all scalar and functional variables used in the model.}

\item{joint_FPCA}{A True/False vector of the same length of the number of functional predictors, indicating whether jointly modeling FPCA for the functional predictors. Default to NULL.}

\item{runStan}{True/False variable for whether to run the Stan program. If False, the function only generates the Stan code and data.}

\item{niter}{Total number of Bayesian iterations.}

\item{nwarmup}{Number of warmup (burnin) iterations for posterior sampling.}

\item{nchain}{Number of chains for posterior sampling. Default to 3.}

\item{ncores}{Number of cores to use when executing the chains in parallel. Default to 1.}

\item{spline_type}{Type of spline basis for modelling the residual process.}

\item{spline_df}{Degrees of freedom for the spline basis for modelling the residual process.}
}
\value{
A list containing:
\item{stanfit}{The Stan fit object.}
\item{spline_basis}{Basis functions used to reconstruct the functional coefficients from posterior samples.}
\item{stancode}{A character string containing the code to fit the Stan model.}
\item{standate}{A list containing the data to fit the Stan model.}
\item{int}{A vector containing posterior samples of the intercept term.}
\item{scalar_coef}{A matrix containing posterior samples of scalar coefficients, where each row is one sample and each column is one variable.}
\item{func_coef}{A list containing posterior samples of functional coefficients. Each element is a matrix, where each row is one sample and each column is one location of the functional domain.}
\item{family}{Distribution of the outcome variable.}
}
\description{
Fit the Bayesian Function-on-Scalar Regression (FOSR) model using Stan.
}
\details{
The Bayesian FOSR model is implemented following the tutorial by Jiang et al., 2025.
The model is specified using the same syntax as in the R mgcv package.
}
\references{
Jiang, Z., Crainiceanu, C., and Cui, E. (2025). Tutorial on Bayesian Functional Regression Using Stan. \emph{Statistics in Medicine}, 44(20-22), e70265.
}
\author{
Erjia Cui \email{ecui@umn.edu}, Ziren Jiang \email{jian0746@umn.edu}
}
\examples{
\dontrun{
## Not run: Stan requires a working C++ toolchain
# Function-on-scalar regression
set.seed(1)
n <- 100; M <- 50
tindex   <- seq(0, 1, length.out = M)
age      <- rnorm(n)
sex      <- rbinom(n, 1, 0.5)
Y_mat    <- outer(age, sin(2 * pi * tindex)) +
            outer(sex, cos(2 * pi * tindex)) +
            matrix(rnorm(n * M, sd = 0.3), nrow = n)
dat       <- data.frame(age = age, sex = sex)
dat$Y_mat <- Y_mat
fit <- fosr_bayes(
  formula = Y_mat ~ age + sex,
  data = dat, spline_type = "bs", spline_df = 10, niter = 2000
)
plot(fit)
}
}

