% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesformula.R
\name{is.func.cov}
\alias{is.func.cov}
\title{A function for deciding whether the object is a functional object with smoothness}
\usage{
is.func.cov(term)
}
\description{
A function for deciding whether the object is a functional object with smoothness
}
