% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayes_plot.R
\name{plot.bfrs}
\alias{plot.bfrs}
\title{Plot the estimated functional coefficients with the corresponding credible interval}
\usage{
plot.bfrs(brfs.fit = NULL, prob = 0.95, include = "both")
}
\description{
Plot the estimated functional coefficients with the corresponding credible interval
}
