% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesreconfuncoef.R
\name{recon_fun_coef}
\alias{recon_fun_coef}
\title{Functions for reconstructing the estimated functional coefficient based on the posterior sampling}
\usage{
recon_fun_coef(basis, fit.samp, func_comp, trans.mat)
}
\description{
The function for reconstructing the SoFR fitted functional coefficient according to Stan posterior sampling
}
