% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesreconfuncoef.R
\name{recon_fun_coef_fosr}
\alias{recon_fun_coef_fosr}
\title{The function for reconstructing the FoSR fitted functional coefficient according to Stan posterior sampling}
\usage{
recon_fun_coef_fosr(basis, b.samp)
}
\description{
The function for reconstructing the FoSR fitted functional coefficient according to Stan posterior sampling
}
