% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesdata.R
\name{refundBayesdata}
\alias{refundBayesdata}
\title{Functions to build the data and code for subsequent Stan modeling}
\usage{
refundBayesdata(
  formula,
  data,
  family,
  func_comp,
  intercept,
  cens = NULL,
  func_parameter = NULL
)
}
\description{
The main function for constructing the data and code used as input for Stan
}
