% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesformula.R
\name{refundBayesformula}
\alias{refundBayesformula}
\title{Functions to generate the formula for subsequent Stan modeling.}
\usage{
refundBayesformula(formula)
}
\description{
The main function for generating the formula as input for other functions
}
