% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesmodel.R
\name{refundBayesmodel}
\alias{refundBayesmodel}
\title{Functions to build the model block of the Stan code.}
\usage{
refundBayesmodel(formula, data, family, func_comp, intercept)
}
\description{
The main function for building the model block for SoFR and functional Cox model.
}
