% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesmodel.R
\name{refundBayesmodel_functional}
\alias{refundBayesmodel_functional}
\title{The main function for building the model block for FoSR.}
\usage{
refundBayesmodel_functional(formula, standata, family, func_comp, intercept)
}
\description{
The main function for building the model block for FoSR.
}
