% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayesmodel.R
\name{set_prior}
\alias{set_prior}
\title{A convenient function for setting the priors.}
\usage{
set_prior(formula, data, family, func_comp, intercept)
}
\description{
A convenient function for setting the priors.
}
