% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofr_bayes.R
\name{sofr_bayes}
\alias{sofr_bayes}
\title{Bayesian Scalar-on-Function Regression}
\usage{
sofr_bayes(
  formula,
  data,
  family = gaussian(),
  joint_FPCA = NULL,
  intercept = TRUE,
  runStan = TRUE,
  niter = 3000,
  nwarmup = 1000,
  nchain = 3,
  ncores = 1
)
}
\arguments{
\item{formula}{Functional regression formula, with the same syntax as that in the R mgcv package.}

\item{data}{A data frame containing data of all scalar and functional variables used in the model.}

\item{family}{Distribution of the outcome variable. Currently support "gaussian" and "binomial".}

\item{joint_FPCA}{A True/False vector of the same length of the number of functional predictors, indicating whether jointly modeling FPCA for the functional predictors. Default to NULL.}

\item{intercept}{True/False variable for whether include an intercept term in the linear predictor. Default to TRUE.}

\item{runStan}{True/False variable for whether to run the Stan program. If False, the function only generates the Stan code and data.}

\item{niter}{Total number of Bayesian iterations.}

\item{nwarmup}{Number of warmup (burnin) iterations for posterior sampling.}

\item{nchain}{Number of chains for posterior sampling. Default to 3.}

\item{ncores}{Number of cores to use when executing the chains in parallel. Default to 1.}
}
\value{
A list containing:
\item{stanfit}{The Stan fit object.}
\item{spline_basis}{Basis functions used to reconstruct the functional coefficients from posterior samples.}
\item{stancode}{A character string containing the code to fit the Stan model.}
\item{standate}{A list containing the data to fit the Stan model.}
\item{int}{A vector containing posterior samples of the intercept term.}
\item{scalar_coef}{A matrix containing posterior samples of scalar coefficients, where each row is one sample and each column is one variable.}
\item{func_coef}{A list containing posterior samples of functional coefficients. Each element is a matrix, where each row is one sample and each column is one location of the functional domain.}
\item{family}{Distribution of the outcome variable.}
}
\description{
Fit the Bayesian Scalar-on-Function Regression (SoFR) model using Stan.
}
\details{
The Bayesian SoFR model is implemented following the tutorial by Jiang et al., 2025.
The model is specified using the same syntax as in the R mgcv package.
}
\references{
Jiang, Z., Crainiceanu, C., and Cui, E. (2025). Tutorial on Bayesian Functional Regression Using Stan. \emph{Statistics in Medicine}, 44(20-22), e70265.
}
\author{
Erjia Cui \email{ecui@umn.edu}, Ziren Jiang \email{jian0746@umn.edu}
}
