% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayes_summary.R
\name{summary.bfrs}
\alias{summary.bfrs}
\title{Generate the summary table for the Bayesian model}
\usage{
summary.bfrs(brfs.fit = NULL, prob = 0.95)
}
\description{
Generate the summary table for the Bayesian model
}
