% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayes_summary.R
\name{summary.refundBayes}
\alias{summary.refundBayes}
\title{Generate the summary table for the Bayesian model}
\usage{
\method{summary}{refundBayes}(object = NULL, ..., prob = 0.95)
}
\arguments{
\item{object}{A fitted object returned by \code{sofr_bayes()}.}

\item{...}{Other parameters}

\item{prob}{Coverage probability for the reported confidence intervals. Defaults to 0.95.}
}
\value{
A list of two objects, the first is the summary table for the estimated scalar coefficients, the second is the plots for the estimated functional coefficients.
}
\description{
Generate the summary table for the Bayesian model
}
