% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundBayes_plot.R
\name{plot.refundBayes}
\alias{plot.refundBayes}
\title{Plot the estimated functional coefficients with the corresponding credible interval(s).}
\usage{
\method{plot}{refundBayes}(x = NULL, ..., prob = 0.95, include = "both")
}
\arguments{
\item{x}{A fitted object returned by \code{sofr_bayes()}.}

\item{...}{Other parameters}

\item{prob}{Coverage probability for the credible interval(s). Defaults to 0.95.}

\item{include}{Type of interval to include. \code{"pointwise"} produces pointwise credible intervals;
\code{"CMA"} produces the CMA credible band; \code{"both"} produces both. Defaults to \code{"both"}.}
}
\value{
A list of \code{ggplot} objects, one for each functional coefficient.
}
\description{
Plot the estimated functional coefficients with the corresponding credible interval(s).
}
